Add MPI integration and write trajs to binary file
Work division between MPI processes is now implemented. The trajectories
are now stored in a binary file so that MPI can do parallel IO (and also
save some storage space). Also the OpenMP code saves the trajectories in
a binary file. The simple plotting script was updated to reflect the
change in output file.